requirements.txt
# LiveKit Agents and Plugins
livekit-agents[google]>=1.3.0
livekit-api>=0.8.0

# Google AI
google-generativeai>=0.8.0

# Additional dependencies
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio>=3.4.3

# Data handling
pydantic>=2.0.0

# Logging and monitoring
structlog>=24.0.0

# Development tools (optional)
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
flake8>=7.0.0
